+++ semantic-conventions/src/opentelemetry/semconv/templating/code.py
@@ -16,12 +16,95 @@ import datetime
 import os.path
 import re
 import typing
+import mistune
 
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
+from opentelemetry.semconv.model.semantic_attribute import TextWithLinks
 from opentelemetry.semconv.model.semantic_convention import SemanticConventionSet
 from opentelemetry.semconv.model.utils import ID_RE
 from opentelemetry.semconv.model.semantic_attribute import Required
+
+def render_markdown(
+    txt: str,
+    html=True,
+    link=None,
+    image=None,
+    emphasis=None,
+    strong=None,
+    inline_html=None,
+    paragraph=None,
+    heading=None,
+    block_code=None,
+    block_quote=None,
+    list=None,
+    list_item=None,
+    code=None,
+):
+    class CustomRender(mistune.HTMLRenderer):
+        def link(self, url, text=None, title=None):
+            if link:
+                return link.format(url, text, title)
+            return super().link(url, text, title) if html else url
+
+        def image(self, src, alt="", title=None):
+            if image:
+                return image.format(src, alt, title)
+            return super().image(src, alt, title) if html else src
+
+        def emphasis(self, text):
+            if emphasis:
+                return emphasis.format(text)
+            return super().emphasis(text) if html else text
+
+        def strong(self, text):
+            if strong:
+                return strong.format(text)
+            return super().strong(text) if html else text
+
+        def inline_html(self, html_text):
+            if inline_html:
+                return inline_html.format(html_text)
+            return super().inline_html(html_text) if html else html_text
+
+        def paragraph(self, text):
+            if paragraph:
+                return paragraph.format(text)
+            return super().paragraph(text) if html else text
+
+        def heading(self, text, level):
+            if heading:
+                return heading.format(text, level)
+            return super().heading(text, level) if html else text
+
+        def block_code(self, code, info=None):
+            if block_code:
+                return block_code.format(code)
+            return super().block_code(code, info) if html else code
+
+        def block_quote(self, text):
+            if block_quote:
+                return block_quote.format(text)
+            return super().block_quote(text)
+
+        def list(self, text, ordered, level, start=None):
+            if list:
+                return list.format(text)
+            return super().list(text, ordered, level, start) if html else text
+
+        def list_item(self, text, level):
+            if list_item:
+                return list_item.format(text)
+            return super().list_item(text, level) if html else text
+
+        def codespan(self, text):
+            if code:
+                return code.format(text)
+            return super().codespan(text) if html else text
+
+    markdown = mistune.create_markdown(renderer=CustomRender())
+    return markdown(txt)
+
 
 def to_doc_brief(doc_string: typing.Optional[str]) -> str:
     if doc_string is None:
@@ -30,6 +113,29 @@ def to_doc_brief(doc_string: typing.Optional[str]) -> str:
     if doc_string.endswith("."):
         return doc_string[:-1]
     return doc_string
+
+
+def to_html_links(doc_string: typing.Optional[typing.Union[str, TextWithLinks]]) -> str:
+    if doc_string is None:
+        return ""
+    if isinstance(doc_string, TextWithLinks):
+        str_list = []
+        for elm in doc_string.parts:
+            if isinstance(elm, str):
+                str_list.append(elm)
+            else:
+                str_list.append('<a href="{}">{}</a>'.format(elm.url, elm.text))
+        doc_string = "".join(str_list)
+    doc_string = doc_string.strip()
+    if doc_string.endswith("."):
+        return doc_string[:-1]
+    return doc_string
+
+
+def regex_replace(text: str, pattern: str, replace: str):
+    # convert standard dollar notation to python
+    replace = re.sub(r"\$", r"\\", replace)
+    return re.sub(pattern, replace, text, 0, re.U)
 
 
 def merge(list: typing.List, elm):
@@ -93,6 +199,9 @@ class CodeRenderer:
         env.filters["to_const_name"] = to_const_name
         env.filters["merge"] = merge
         env.filters["to_camelcase"] = to_camelcase
+        env.filters["to_html_links"] = to_html_links
+        env.filters["regex_replace"] = regex_replace
+        env.filters["render_markdown"] = render_markdown
 
     @staticmethod
     def prefix_output_file(file_name, pattern, semconv):