+++ semantic-conventions/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Please update the changelog as part of any significant pull request.
 
+## v0.4.0
+
+- Add stability fields. (#35)
+- Add Markdown render for code generation.
+
 ## v0.3.1
 
 - Fix markdown generator for int enums. (#36)